day ant diary fix

+++ app/Http/Controllers/DayApiController.php
@@ -10,23 +10,38 @@ class DayApiController extends Controller
 {
     public function index()
     {
-        return Day::all();
+        $isGuest = auth()->guest();
+        
+        if(!$isGuest)
+        {
+            $user_id = auth()->user()->id;
+            return Day::where('user_id', $user_id)->get();
+        }
+        else
+        {
+            return response()->json(["message" => "Unauthorized"], 401);
+        }
     }
 
     public function store()
     {
         $isGuest = auth()->guest();
 
-        if (!$isGuest) {
+        if (!$isGuest) 
+        {
+            $user_id = auth()->user()->id;
 
-            if (Trips::where('id', request('trip_id'))->exists()) {
-                return Day::create(['day_number' => request('day_number'), 'trip_id' => request('trip_id'), 'budget' => request('budget'), 'note' => request('note'),]);
+            if (Trips::where('id', request('trip_id'))->exists()) 
+            {
+                return Day::create(['day_number' => request('day_number'), 'trip_id' => request('trip_id'), 'budget' => request('budget'), 'note' => request('note'), 'user_id' => $user_id]);
             } 
-            else {
+            else 
+            {
                 return response()->json(["message" => "Trip not found"], 404);
             }
         } 
-        else {
+        else 
+        {
             return response()
                 ->json(["message" => "Unauthorized"], 401);
         }
@@ -34,43 +49,70 @@ class DayApiController extends Controller
 
     public function update(Request $request, $id)
     {
-
         $isGuest = auth()->guest();
 
-        if (!$isGuest) {
+        if (!$isGuest) 
+        {
             if (Day::where('id', $id)->exists()) {
-                $day = Day::find($id);
-                $day->day_number = $day->day_number;
-                $day->trip_id = $day->trip_id;
-                $day->budget = is_null($request->budget) ? $day->budget : $request->budget;
-                $day->note = is_null($request->note) ? $day->note : $request->note;
-                $day->save();
 
-                return response()
-                    ->json(["message" => "Day updated successfully"], 200);
-            } 
-            else {
+                $day = Day::find($id);
+                $user_id=auth()->user()->id;
+
+                if($user_id == $day->user_id)
+                {
+                    $day->day_number = is_null($request->day_number) ? $day->day_number : $request->day_number;
+                    $day->user_id = $day->user_id;
+                    $day->trip_id = $day->trip_id;
+                    $day->budget = is_null($request->budget) ? $day->budget : $request->budget;
+                    $day->note = is_null($request->note) ? $day->note : $request->note;
+                    $day->save();
+
+                    return response()
+                        ->json(["message" => "Day updated successfully"], 200);
+                } 
+                else
+                {
+                    return response()->json(["message" => "Unauthorized"], 401);
+                }
+            }
+            else 
+            {
                 return response()
                     ->json(["message" => "Day not found"], 404);
             }
+                
         } 
-        else {
+        else 
+        {
             return response()
                 ->json(["message" => "Unauthorized"], 401);
         }
+
     }
 
     public function destroy($id)
     {
-
         $isGuest = auth()->guest();
-        if (!$isGuest) {
-            if (Day::where('id', $id)->exists()) {
-                $day = Day::find($id);
-                $day->delete();
 
-                return response()
-                    ->json(["message" => "Day deleted"], 202);
+        if (!$isGuest) 
+        {
+            $user_id = auth()->user()->id;
+
+            if (Day::where('id', $id)->exists()) 
+            {
+                $day = Day::find($id);
+
+                if ($user_id == $day->user_id) 
+                {
+                    $day->delete();
+
+                    return response()
+                        ->json(["message" => "Day deleted"], 202);
+                } 
+                else {
+                    return response()
+                        ->json(["message" => "Unauthorized"], 401);
+                }
             } 
             else {
                 return response()
@@ -89,15 +131,25 @@ class DayApiController extends Controller
 
         if (!$isGuest) 
         {            
-            if(Day::where('id', $id)->exists())
+            $user_id = auth()->user()->id;
+            $day = Day::find($id);
+            if ($user_id == $day->user_id) 
             {
-                return Day::find($id);
+                if(Day::where('id', $id)->exists())
+                {
+                    return Day::find($id);
+                }
+                else
+                {
+                    return response()->json([
+                        "message" => "Day not found"
+                    ], 404);
+                }
             }
-            else
+            else 
             {
-                return response()->json([
-                    "message" => "Day not found"
-                ], 404);
+                return response()
+                    ->json(["message" => "Unauthorized"], 401);
             }
         } 
         else
